This One Weird Trick Speeds Up Startup By 50%
This is a horrible horrible hack.  If for no other reason that (without
EJSON.clone around) it leaves internal pieces of catalog.official and
catalog.complete intertwined.

But! It does mean that every single tool startup now only has to read
and parse packages.data.json once instead of twice.  Which speeds up
'meteor --version' from 1 second to 0.5.

We'll solve this for real with the sqlite refactoring. But this is fast
and easy for now.