Use ChildProcess kill method to reload CSS
Using the method instead of process.kill means that we don't send the
signal if we already know the process is dead and that we don't throw on
ESRCH (ie, the kernel telling us the process is gone).  (If the process
has already died we will soon get around to doing a different kind of
restart.)

Fixes #2550.