Use `project.setRootDir` instead of `reload` in stats test.
As of 3e64bd78, this usage of `reload` no longer works (if we set a
project's root dir to a fresh app, and then run the app which computes
versions.json, and then call `reload`, we'll end up with the project's
dependencies being empty). Maybe this is a bug in `reload`, but it
probably doesn't matter for any usage other than this test, and maybe
this test was misusing `reload` in the first place.