Fix appcache usage of WebApp.clientHash
In commit 71966f40 `WebApp.clientHash` changed from a string to a
function.  This fixes the appcache package to call the new function.

With the change to a function being called, the client hash is no
longer cached and the client hash will be recalculated on each load of
the app manifest file.

I presume the client hash could be cached by the appcache package
because it always uses the default architecture ("web.browser"), but I
didn't make that addition because I was unsure.