Update 'add cordova plugins' test to reflect reality.
If we fail to install one plugin, then we uninstall all plugins to avoid
awakening the Cordova monsters. Previously, this test was passing
because we were passing a `--settings` argument, which was allowing the
facebookconnect plugin to install, but the build to fail to other,
not-understood reasons. Now that we've removed `--settings`, the plugin
fails to install, which means that all plugins get uninstalled.

Test passes now.