Try to not include ctl program in apps that don't need it.
This may be close to working. It works fine when ctl is not included,
but when ctl is added to the app to get it to deploy to galaxy, the
app doesn't start because there are 2 main functions, because the ctl
code is included in the server program as well.

Parking this for now, until I can get someone who understands this
to look at it.