Don't automatically put ?close at the end of Meteor Developer Accounts URLs
MDA stuff configured with the new tool does not use ?close, and the one configured with
the old tool, does. We are not going to automatically try contacting the ?close anymore,
since that doesn't work in the future; for backwards compatibility, we went through package
server & stat server and added the close-less urls for the new tool to use.