Print in console when we are restarting the server
(…due to a server-side code change)

Restarting the server can take a while (30 seconds for one particular app), so it’s nice to at least print when we start doing it.  Soon there will be progress displays during it, too (when we add more of them to `meteor run`).

We lose the opportunity to coalesce server restarted messages, as in `Meteor server restarted (x2)`, but adding any sort of progress display also does that.