Display progress when restarting the server
Previously, all progress display was turned off after the app’s initial start, so no progress would be displayed on restart due to server code change.

Now, we enable/disable progress display from inside the loop that re-runs the app on change.

Remove the “suppressDisplay” approach to hiding the spinner when there’s an error building the app, because it didn’t work on rebuilds, only the first build.  Better to just call enableProgressDisplay(false) and hope for the best like we do elsewhere.

Conflicts:
	tools/progress.js