Fix “Uploading” progress bar on meteor deploy
For some reason, super-jobs of the actual upload job have a `.current` progress but no `.end` value, and then they are chosen for progress display instead of the actual upload.  With this change, we don’t eagerly choose a job for display if it has a `.current` unless it also has a `.end`.

It’s probably also worth examining what is going on with the various values of `.current` of the super-jobs.