Remove release methods from catalog.complete
They are unused (and were accidentally introduced in 0.9.4 with the
sqlite branch).  The only thing you're allowed to query for releases is
the official catalog, since the complete catalog doesn't actually
augment releases with anything.

(If we later implement some sort of "local release", we should probably
also introduce a new release-specific interface that can be backed
either by sqlite or by sqlite-plus-local, analogous to catalog.  There's
no real reason that it should be the same interface as catalog.)