Isopack.saveToPath doesn't always need a capture
It should only need a capture if you need to include
buildinfo. Specifically, the use in tropohouse for package merging
should not be able to fail with a buildmessage.

Failures to statically compile packages for the tool now just crash the
tool, just like uniload failures. This should never happen unless you're
working from a checkout and have broken a core package.