Ignore .build.* dirs when building isopackets
Neither read from them nor write to them.

Yes, this means that when running from a checkout, rebuilding an
isopacket (eg because you modified ddp or something) is a little
slower. But it's more correct --- having a single cache in the source
tree that's shared between apps (which run in catalog.complete mode) and
isopackets has some serious correctness issues.  And the common case
where no isopackets have to be rebuilt is still fast (./meteor --version
takes about 0.24 seconds).