Handle plugin dependencies in constraint solver
This bug was discovered when trying to run an app which depended only on
'test-in-browser': templating would not build because spacebars-compiler
was only a dependency of the plugin in templating and had not been
built.

The "plugin" arch detected here didn't even exist before the
isopack-cache branch. So 1.0 probably has bugs where constraints set in
plugin dependencies are ignored.

**This code doesn’t do anything without glasser’s isopack-cache branch.  Cherry-picking it now to avoid conflicts.**