Create PackageGraph class
PackageGraph is a simple, JSONable format designed to sit between the Catalog and the generation of unit-unit constraints.  It’s not done; it will get an “addDependency” method, for example, and ways to read the data.  Then we will generate a UnitGraph which expands out all the package#arch units and generates some extra constraints according to our special rules about archs.  The UnitGraph will be input to the solver.

Having PackageGraph will make it easier to generate test cases from real failures (just dump the PackageGraph).