Factor data out of Resolver into CatalogCache
This is a behavior-preserving change.  It’s a first step towards explicitly separating the dependency/constraint info we load from the catalog and the info we pass to the solver.  The idea is you can populate a CatalogCache either from the catalog or from JSON (for tests).  Then, you can use the CatalogCache to either run the old solver (which uses mori, interns constraint objects, etc.) or the new solver (which does something else).