Make --get-ready do something useful
This command was changed in 0.9.0 to do something of questionable use,
and was temporarily unimplemented as part of work on #3006.  It now does
something relatively simple to describe: It prepares your current app
for build, and in addition, it downloads EVERY package in your current
release and builds EVERY available local package (including those which
you are not using in your app).