Sometimes we construct paths w/o them being absolute.
For example: files.pathJoin('/something/', '/something/else') should
result into '/something/something/else' and here the caller doesn't
care if the passed arguments look like absolute paths or not. It can
even be a middle part of some larger path. We don't care and shouldn't
enforce it on files.pathJoin