Don't force external links to open in a new tab
Overriding user preference is a bad idea from a usability point of view. Just allow people to decide for themselves whether they want to navigate away to another page or open the link in a new tab using the controls provided them by the browser (eg. cmd-click, or the options in the right click menu).

This also fixes the situation where someone might want to open an external link in the same tab, which is currently made impossible.