Add Logic.and; fix formula generation
We weren’t generating formulas when they were used by other formulas.  When we generate a clause for a formula, like “A v B v -$or1”, we need to only suppress formula generation for the -$or1 term, not the entire clause.  This matters when A and B are formula vars (eg “$and1 v $and2 v -$or1”).